Refactor tests into separate subproject
