--- core/src/commonTest/kotlin/komem.litmus/testsuite/CustomTests.kt
@@ -1,9 +1,0 @@
-package komem.litmus.testsuite
-
-import komem.litmus.infra.run
-import kotlin.test.Test
-
-class CustomTests {
-    @Test
-    fun mpNoDrf() = MPNoDRF.run()
-}
