+++ core/src/nativeMain/kotlin/komem.litmus/barriers/CinteropSpinBarrier.kt
@@ -3,6 +3,7 @@ package komem.litmus.barriers
 import barrier.CSpinBarrier
 import barrier.barrier_wait
 import barrier.create_barrier
+import komem.litmus.Barrier
 import kotlinx.cinterop.CPointer
 import kotlinx.cinterop.ExperimentalForeignApi
