+++ jcstress-wrapper/build.gradle.kts
@@ -7,7 +7,12 @@ application {
     mainClass = "MainKt"
 }
 
+kotlin {
+    jvmToolchain(8)
+}
+
 dependencies {
     implementation(project(":core"))
+    implementation(project(":testsuite"))
     implementation(kotlin("reflect"))
 }
