+++ settings.gradle.kts
@@ -4,3 +4,4 @@ include(":core")
 include(":codegen")
 include(":jcstress-wrapper")
 include(":cli")
+include(":testsuite")
