+++ testsuite/src/commonMain/kotlin/komem/litmus/LitmusTestExtensions.kt
@@ -1,0 +1,6 @@
+package komem.litmus
+
+import komem.litmus.generated.LitmusTestRegistry
+
+val LitmusTest<*>.name get() = LitmusTestRegistry.resolveName(this)
+val LitmusTest<*>.javaClassName get() = name.replace('.', '_')
